pyproject.toml
[tool.poetry]
name = "kafka-playground"
version = "0.1.0"
description = ""
authors = ["Jose Hidalgo <joseangel1196@hotmail.com>"]

[tool.poetry.dependencies]
python = "3.8.13"
kafka-python = "2.0.2"
pydantic = "^1.10.7"
faker = "^18.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
